+++ Minio/MinioClientExtensions.cs
@@ -276,27 +276,6 @@ public static class MinioClientExtensions
         return minioClient;
     }
 
-    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings",
-        Justification = "This is done in the interface. String is provided here for convenience")]
-    public static Task<HttpResponseMessage> WrapperGetAsync(this IMinioClient minioClient, string url)
-    {
-        if (minioClient is null) throw new ArgumentNullException(nameof(minioClient));
-
-        return minioClient.WrapperGetAsync(new Uri(url));
-    }
-
-    /// <summary>
-    ///     Runs httpClient's PutObjectAsync method
-    /// </summary>
-    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings",
-        Justification = "This is done in the interface. String is provided here for convenience")]
-    public static Task WrapperPutAsync(this IMinioClient minioClient, string url, StreamContent strm)
-    {
-        if (minioClient is null) throw new ArgumentNullException(nameof(minioClient));
-
-        return minioClient.WrapperPutAsync(new Uri(url), strm);
-    }
-
     /// <summary>
     ///     Sets app version and name. Used for constructing User-Agent header in all HTTP requests
     /// </summary>
