+++ Minio/RequestExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Minio.Credentials;
+﻿using System.Diagnostics.CodeAnalysis;
+using Minio.Credentials;
 using Minio.DataModel;
 using Minio.DataModel.Args;
 using Minio.DataModel.Result;
@@ -7,8 +8,29 @@ using Minio.Helper;
 
 namespace Minio;
 
-internal static class RequestExtensions
+public static class RequestExtensions
 {
+    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings",
+        Justification = "This is done in the interface. String is provided here for convenience")]
+    public static Task<HttpResponseMessage> WrapperGetAsync(this IMinioClient minioClient, string url)
+    {
+        if (minioClient is null) throw new ArgumentNullException(nameof(minioClient));
+
+        return minioClient.WrapperGetAsync(new Uri(url));
+    }
+
+    /// <summary>
+    ///     Runs httpClient's PutObjectAsync method
+    /// </summary>
+    [SuppressMessage("Design", "CA1054:URI-like parameters should not be strings",
+        Justification = "This is done in the interface. String is provided here for convenience")]
+    public static Task WrapperPutAsync(this IMinioClient minioClient, string url, StreamContent strm)
+    {
+        if (minioClient is null) throw new ArgumentNullException(nameof(minioClient));
+
+        return minioClient.WrapperPutAsync(new Uri(url), strm);
+    }
+
     /// <summary>
     ///     Actual doer that executes the request on the server
     /// </summary>
